requirements.txt
Django==1.11.5
django-ckeditor==5.3.0
django-js-asset==0.1.1
freeze==1.0.10
olefile==0.44
Pillow==4.2.1
pytz==2017.2
six==1.11.0